refactor: refactor formula menu items using composition
