+++ components/formula/FormulaMenu/components/Favorite/index.tsx
@@ -1,38 +1,24 @@
-import {
-  ActionIcon,
-  Menu,
-  MenuItemProps,
-  Text,
-  useMantineTheme,
-} from "@mantine/core";
 import { IconDeviceFloppy } from "@tabler/icons-react";
+import { FormulaMenuItem } from "components/formula/FormulaMenu/components/Item";
+import { FormulaMenuItemProps } from "components/formula/FormulaMenu/components/Item/models";
 import { useFavoritesStore } from "modules/favorites";
 import { selectAddFavorite } from "modules/favorites/selectors";
 import { FormulaComputable } from "modules/formulas/models";
 import React, { FC } from "react";
 
-interface Props extends MenuItemProps {
+interface Props extends FormulaMenuItemProps {
   computable: FormulaComputable;
 }
 
 export const FormulaMenuFavorite: FC<Props> = ({ computable, ...props }) => {
-  const theme = useMantineTheme();
   const addFavorite = useFavoritesStore(selectAddFavorite);
   const onAddFavorite = () => {
     addFavorite(computable);
   };
 
   return (
-    <Menu.Item
-      icon={
-        <ActionIcon size={"xs"} variant={"transparent"} component={"span"}>
-          <IconDeviceFloppy size={theme.spacing.sm} />
-        </ActionIcon>
-      }
-      onClick={onAddFavorite}
-      {...props}
-    >
-      <Text size={"xs"}>Save</Text>
-    </Menu.Item>
+    <FormulaMenuItem onClick={onAddFavorite} icon={IconDeviceFloppy} {...props}>
+      Save
+    </FormulaMenuItem>
   );
 };
