+++ components/formula/FormulaMenu/components/Item/models.ts
@@ -1,0 +1,12 @@
+import { MenuItemProps } from "@mantine/core";
+import { Icon } from "@tabler/icons-react";
+import { ComponentPropsWithoutRef } from "react";
+
+type ButtonProps = ComponentPropsWithoutRef<"button">;
+
+export interface FormulaMenuItemProps
+  extends Omit<MenuItemProps, "icon">,
+    Omit<ButtonProps, keyof MenuItemProps> {
+  icon?: Icon;
+  disabled?: boolean;
+}
